Fix compile warnings (#1762)
* Fix comparison between signed and unsigned integer expressions
* Remove unused variable

+++ torchaudio/csrc/pybind/sox/effects.cpp
@@ -54,7 +54,7 @@ auto apply_effects_fileobj(
     // end up retrieving the static variable defined in `_torchaudio`, which is
     // not correct.
     const auto bufsiz = get_buffer_size();
-    const size_t kDefaultCapacityInBytes = 256;
+    const int64_t kDefaultCapacityInBytes = 256;
     return (bufsiz > kDefaultCapacityInBytes) ? bufsiz
                                               : kDefaultCapacityInBytes;
   }();
